# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RepoLens API",
    "description": "Deterministic, evidence-backed analysis of a public GitHub repository at an exact commit.",
    "license": {
      "name": "LicenseRef-Proprietary",
      "identifier": "LicenseRef-Proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/analyses": {
      "post": {
        "tags": [
          "analyses"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted; the analysis is queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            }
          },
          "400": {
            "description": "The URL is not a public GitHub repository",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "The analysis store is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "operationId": "read",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "description": "Analysis identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current state of the analysis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            }
          },
          "404": {
            "description": "No such analysis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "The analysis store is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/report": {
      "get": {
        "tags": [
          "analyses"
        ],
        "operationId": "read_report",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "description": "Analysis identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The completed report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "404": {
            "description": "No report for that analysis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "The analysis store is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Analysis": {
        "type": "object",
        "description": "One analysis run.",
        "required": [
          "id",
          "repository",
          "commit_sha",
          "state",
          "execution",
          "retry",
          "created_at",
          "updated_at",
          "report_available"
        ],
        "properties": {
          "commit_sha": {
            "type": [
              "string",
              "null"
            ],
            "description": "The exact commit, once resolved.\n\n**Null during `QUEUED` and `RESOLVING`** — there genuinely is no commit\nyet. Required-but-nullable rather than optional, so a consumer cannot\nforget the case: the field is always present, its value is not."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the analysis was created."
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError",
                "description": "Why it failed. Present only in a failed state."
              }
            ]
          },
          "execution": {
            "$ref": "#/components/schemas/ExecutionMetadata",
            "description": "Scheduling facts, kept out of `state`."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable identifier. `UUIDv7`: time-ordered for index locality, and with 74\nrandom bits it remains unguessable, which is what allows anonymous\nprogress viewing by URL."
          },
          "poll_after_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "How long the client should wait before polling again.\n\nServer-supplied so the interval can widen as an analysis ages, and so a\nhardcoded frontend value cannot multiply cold starts and cost. Absent in\nterminal states — there is nothing left to poll for.",
            "minimum": 0
          },
          "report_available": {
            "type": "boolean",
            "description": "Whether `GET /reports/{id}` will return a report.\n\nExplicit rather than `state == COMPLETED`, because report availability\nand analysis completion are separate facts once reports are retained,\npruned, or regenerated under a newer ruleset."
          },
          "repository": {
            "$ref": "#/components/schemas/RepositoryIdentity",
            "description": "Owner and name, known from creation."
          },
          "retry": {
            "$ref": "#/components/schemas/RetryPolicy",
            "description": "Server's decision on whether a retry is currently permitted."
          },
          "state": {
            "$ref": "#/components/schemas/AnalysisState",
            "description": "Where the analysis has reached."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When it last changed state. Distinct from `created_at` so a UI can show\n\"stuck for 20 minutes\" rather than only \"started 20 minutes ago\"."
          }
        }
      },
      "AnalysisState": {
        "type": "string",
        "description": "Where an analysis has reached.\n\nOrdered as the work actually proceeds, so a UI can render a checklist by\nposition without a second table mapping states to steps.\n\nInfrastructure state is deliberately absent — there is no `TRIGGERING` or\n`WAITING_FOR_WORKER`. Whether a Cloud Run Job execution was accepted is a\nproperty of the *execution*, not of the analysis, and mixing them would mean\nevery consumer of this enum had to learn how the work is scheduled. See\n[`ExecutionMetadata`].",
        "enum": [
          "QUEUED",
          "RESOLVING",
          "COLLECTING",
          "ANALYZING",
          "BUILDING_REPORT",
          "COMPLETED",
          "FAILED_RETRIABLE",
          "FAILED_PERMANENT"
        ]
      },
      "ApiError": {
        "type": "object",
        "description": "A failure, as the browser receives it.\n\nFields are private and deserialization is validated, so\n`ANALYZER_FAILED_PERMANENT` carrying a 900-second countdown cannot be\nconstructed *or* parsed. A safe constructor alone would not achieve that:\nderived `Deserialize` would still accept the combination off the wire, and a\npublic field would still accept it in Rust.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable machine code. Switch on this, never on `message`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation. Safe to display, and deliberately free of\ninternal identifiers, hostnames, and credentials — this string crosses\ninto a browser."
          },
          "retry_after_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "How long to wait before retrying, when that is actually knowable.\n\nAbsent rather than zero when unknown: a UI that renders \"retry in 0s\"\nfrom a missing value is worse than one that renders no countdown at all.",
            "minimum": 0
          }
        }
      },
...
      "Confidence": {
        "type": "string",
        "description": "Strength of the evidence. Independent of [`Severity`].",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH"
        ]
      },
      "CreateAnalysisRequest": {
        "type": "object",
        "description": "What a caller submits to start an analysis.",
        "required": [
          "repository_url"
        ],
        "properties": {
          "repository_url": {
            "type": "string",
            "description": "A public GitHub repository URL, e.g. `https://github.com/rust-lang/crates.io`.\n\nA URL rather than separate owner and name fields, because a URL is what a\nperson has in their clipboard. Parsing it is our job, not theirs."
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "Why a request or an analysis failed.\n\n`SCREAMING_SNAKE_CASE` per the settled convention (#14). The set is closed on\npurpose — see [`super::UNKNOWN_VARIANT_POLICY`] for what the frontend does\nwhen a future backend adds one it has never seen.",
        "enum": [
          "INVALID_REPOSITORY_URL",
          "REPOSITORY_NOT_FOUND",
          "REPOSITORY_INACCESSIBLE",
          "REPOSITORY_ARCHIVED",
          "REPOSITORY_TOO_LARGE",
          "RATE_LIMITED",
          "WORKER_FAILED_RETRIABLE",
          "ANALYZER_FAILED_PERMANENT"
        ]
      },
...
      "ExecutionMetadata": {
        "type": "object",
        "description": "Scheduling facts about an analysis.",
        "required": [
          "trigger_status"
        ],
        "properties": {
          "execution_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Runner-assigned execution identifier, for correlating logs."
          },
          "trigger_status": {
            "$ref": "#/components/schemas/TriggerStatus",
            "description": "Whether the work was successfully handed to a runner."
          },
          "triggered_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the trigger was attempted."
          }
        }
      },
      "Finding": {
        "type": "object",
        "description": "One conclusion, with everything needed to check it.",
        "required": [
          "id",
          "rule_id",
          "ruleset_version",
          "category",
          "state",
          "severity",
          "confidence",
          "title",
          "explanation",
          "evidence",
          "limitations"
        ],
        "properties": {
          "category": {
            "$ref": "#/components/schemas/FindingCategory",
            "description": "Section this belongs to."
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence",
            "description": "Evidence strength. **Never merged with `severity`.**"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "description": "Facts supporting the conclusion. May be empty for `UNABLE_TO_VERIFY`,\nwhich is precisely the case where there is nothing to show."
          },
          "explanation": {
            "type": "string",
            "description": "Prose explanation, including why it matters."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable identifier within this report."
          },
          "limitations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Limitation"
            },
            "description": "What this finding does not establish."
          },
          "recommended_action": {
            "type": [
              "string",
              "null"
            ],
            "description": "Suggested next step. Absent when the honest answer is \"nothing to do\"."
          },
          "rule_id": {
            "type": "string",
            "description": "Which rule produced this, e.g. `rust.workspace.detected`."
          },
          "ruleset_version": {
            "type": "string",
            "description": "Ruleset that produced it. Carried per-finding as well as per-report so a\nstored finding stays interpretable after the report is regenerated."
          },
          "severity": {
            "$ref": "#/components/schemas/Severity",
            "description": "Impact if valid. **Never merged with `confidence`.**"
          },
          "state": {
            "$ref": "#/components/schemas/FindingState",
            "description": "What the analyzer concluded."
          },
          "title": {
            "type": "string",
            "description": "One-line summary."
          }
        }
      },
...
      "Limitation": {
        "type": "object",
        "description": "Something the analyzer could not establish.",
        "required": [
          "code",
          "explanation"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable code, so the UI can group and explain limitations consistently."
          },
          "explanation": {
            "type": "string",
            "description": "What could not be established, and why."
          }
        }
      },
      "LineCountSummary": {
        "type": "object",
        "description": "Repository composition and line counts.\n\nMeasures composition, **not** productivity or quality. The report says so\nvisibly, because this is the easiest section to misread as a score.",
        "required": [
          "counter",
          "counter_version",
          "exclusion_policy_version",
          "total_files",
          "total_lines",
          "code_lines",
          "comment_lines",
          "blank_lines",
          "languages",
          "areas",
          "exclusions",
          "roles",
          "largest_files",
          "unclassified_files"
        ],
        "properties": {
          "areas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AreaLineCount"
            },
            "description": "Per-area breakdown, server-ordered."
          },
          "blank_lines": {
            "type": "integer",
            "format": "int64",
            "description": "Blank lines.",
            "minimum": 0
          },
          "code_lines": {
            "type": "integer",
            "format": "int64",
            "description": "Lines of code.",
            "minimum": 0
          },
          "comment_lines": {
            "type": "integer",
            "format": "int64",
            "description": "Comment lines.",
            "minimum": 0
          },
          "counter": {
            "type": "string",
            "description": "Counting tool, e.g. `tokei`."
          },
          "counter_version": {
            "type": "string",
            "description": "Exact counter version — different versions count differently, so this is\npart of what makes a count reproducible."
          },
          "exclusion_policy_version": {
            "type": "string",
            "description": "Version of the exclusion policy applied."
          },
          "exclusions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompositionExclusion"
            },
            "description": "What was left out, and why."
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LanguageLineCount"
            },
            "description": "Per-language breakdown, server-ordered."
          },
          "largest_files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LargestSourceFile"
            },
            "description": "Largest files by line count, server-ordered, descending.\n\nBounded in both directions by [`LargestSourceFiles`]: a producer cannot\nconstruct an over-long list, and an over-long one cannot be parsed.",
            "maxItems": 10
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleLineCount"
            },
            "description": "Breakdown by role, server-ordered.\n\nPresent so the report can show what proportion of the repository is\nproduction code without implying a judgement about it."
          },
          "total_files": {
            "type": "integer",
            "format": "int64",
            "description": "Files counted.",
            "minimum": 0
          },
          "total_lines": {
            "type": "integer",
            "format": "int64",
            "description": "All physical lines.",
            "minimum": 0
          },
          "unclassified_files": {
            "type": "integer",
            "format": "int64",
            "description": "Files the policy could not classify. Reported rather than silently\nfolded into a bucket.",
            "minimum": 0
          }
        }
      },
...
      "Report": {
        "type": "object",
        "description": "A complete report for one repository at one commit.",
        "required": [
          "analysis_id",
          "repository",
          "commit_sha",
          "tree_sha",
          "analyzer_version",
          "ruleset_version",
          "completed_at",
          "overview",
          "findings",
          "composition",
          "limitations"
        ],
        "properties": {
          "analysis_id": {
            "type": "string",
            "format": "uuid",
            "description": "Analysis that produced it."
          },
          "analyzer_version": {
            "type": "string",
            "description": "Analyzer version that produced this report. First-class, not a footnote."
          },
          "commit_sha": {
            "type": "string",
            "description": "Exact commit. Non-null here, unlike on the analysis: a report cannot\nexist without a resolved commit."
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the analysis completed."
          },
          "composition": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LineCountSummary",
                "description": "Composition and line counts.\n\n**Null when counting did not happen** — most often because extraction\nexceeded its configured limit. That is a designed `UNABLE_TO_VERIFY`\noutcome with a matching entry in `limitations`, not an error and not a\nzero. Required-but-nullable so a consumer must handle it."
              }
            ]
          },
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "description": "All findings, in a **server-decided order**.\n\nOrdering is part of the contract. A report that listed findings\ndifferently on each load would contradict the determinism it claims,\nand no client-side sort can restore an order the server never fixed."
          },
          "limitations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Limitation"
            },
            "description": "What this report as a whole does not establish.\n\nReport-level, not merely per-finding, so \"absence of evidence is not\nevidence of absence\" stays visible in the overview rather than buried\ninside an expanded finding nobody opened."
          },
          "overview": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OverviewStatement"
            },
            "description": "Evidence-backed summary statements."
          },
          "repository": {
            "$ref": "#/components/schemas/RepositoryIdentity",
            "description": "Repository analyzed."
          },
          "ruleset_version": {
            "type": "string",
            "description": "Ruleset version evaluated. First-class, not a footnote."
          },
          "tree_sha": {
            "type": "string",
            "description": "Root tree the collectors walked. Part of the reproducibility key, since\ntwo commits sharing a tree yield identical evidence."
          }
        }
      },
      "RepositoryIdentity": {
        "type": "object",
        "description": "Repository identity, available from the moment an analysis is created.\n\nPresent before `commit_sha` exists, which is what lets the header render\n`owner/name` immediately instead of a blank space that looks broken.",
        "required": [
          "owner",
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Repository name, without the owner prefix."
          },
          "owner": {
            "type": "string",
            "description": "User or organization."
          }
        }
      },
      "RetryPolicy": {
        "type": "object",
        "description": "Whether the caller may retry, decided by the server.\n\nNever inferred from the state name. `FAILED_RETRIABLE` describes the *kind*\nof failure; whether a retry is permitted also depends on how many attempts\nhave already been spent and whether the work is still claimable — facts only\nthe server holds. A frontend that derived this would offer a button that\ndoes nothing.",
        "required": [
          "allowed"
        ],
        "properties": {
          "allowed": {
            "type": "boolean",
            "description": "Whether a retry request would be accepted right now."
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why not, when `allowed` is false. Displayed verbatim, so it explains\nrather than merely denies."
          }
        }
      },
...
      "TriggerStatus": {
        "type": "string",
        "description": "Whether the scheduler accepted the work.\n\nSeparate from [`AnalysisState`] because they fail independently: an analysis\ncan be `QUEUED` with the trigger *succeeded* (normal, waiting for a worker)\nor `QUEUED` with the trigger *failed* (stuck, and nothing will ever pick it\nup). Those look identical without this, and the second one is the outage.",
        "enum": [
          "PENDING",
          "SUCCEEDED",
          "FAILED"
        ]
      }
    }
  }
}